Add files via upload
Completed my algorithm. Altered it to change the city add order based on if the remaining cities were within the polygon path of already captured cities or not. Accuracy of less than 8% for small n. When inputting large n, it is already clear it is not the optimal path. May attempt to utilize clustering algorithms to find clusters, obtain the Center of Mass of those clusters, and further optimize based on how large n could get.